# Conflicts:
#	Gage-master/.idea/caches/build_file_checksums.ser
#	Gage-master/.idea/caches/gradle_models.ser
#	Gage-master/app/src/main/java/uk/ac/qub/eeecs/game/matchAttax/screens/OptionsScreen.java

+++ Gage-master/app/src/main/java/uk/ac/qub/eeecs/game/matchAttax/player/PlayerAI.java
@@ -32,7 +32,7 @@ public class PlayerAI extends Player {
                 }
                 else{
                     if (abs(scoreDiff - card.getOverallValue()) < closestToDiff || closestToDiff == -1){
-                        closestToDiff = abs(scoreDiff - card.getOverallValue());
+                        closestToDiff = abs(scoreDiff - (int)card.getOverallValue());
                         cardToPlay = card;
                     }
                 }
